Move plan types into `cloud_api_types`
